+++ CHANGES.rst
@@ -64,6 +64,8 @@ Internal Changes
 
 - Update Python & JavaScript dependencies (:pr:`5726`, :pr:`5752`)
 - Add support for the watchfiles live reloader (:pr:`5732`)
+- Add an endpoint to allow resetting the state of an accepted editable into "ready to
+  review" (:pr:`5758`)
 
 
 Version 3.2.3